[Enhance] Show brokers' hostname (#5522)
Show brokers' hostname in table to keep the same with show proc for backends and frontends.
