+++ fe/fe-core/src/main/java/org/apache/doris/catalog/BrokerMgr.java
@@ -27,6 +27,7 @@ import org.apache.doris.common.io.Writable;
 import org.apache.doris.common.proc.BaseProcResult;
 import org.apache.doris.common.proc.ProcNodeInterface;
 import org.apache.doris.common.proc.ProcResult;
+import org.apache.doris.common.util.NetUtils;
 import org.apache.doris.common.util.TimeUtils;
 
 import com.google.common.collect.ArrayListMultimap;
@@ -48,9 +49,11 @@ import java.util.concurrent.locks.ReentrantLock;
  */
 public class BrokerMgr {
     public static final ImmutableList<String> BROKER_PROC_NODE_TITLE_NAMES = new ImmutableList.Builder<String>()
-            .add("Name").add("IP").add("Port").add("Alive")
+            .add("Name").add("IP").add("HostName").add("Port").add("Alive")
             .add("LastStartTime").add("LastUpdateTime").add("ErrMsg")
             .build();
+
+    public static final int HOSTNAME_INDEX = 2;
 
     // we need IP to find the co-location broker.
     // { BrokerName -> { IP -> [FsBroker] } }
@@ -341,6 +344,7 @@ public class BrokerMgr {
                         List<String> row = Lists.newArrayList();
                         row.add(brokerName);
                         row.add(broker.ip);
+                        row.add(NetUtils.getHostnameByIp(broker.ip));
                         row.add(String.valueOf(broker.port));
                         row.add(String.valueOf(broker.isAlive));
                         row.add(TimeUtils.longToTimeString(broker.lastStartTime));
